requirements.txt
scrapy
beautifulsoup4
lxml
streamlit
apscheduler
python-dotenv
pytest
